Merges TcpTracker and PortTracker
